Use text input for Canvas access token
When asking for the user's Canvas access token, I initially used a
password input field. However, as discussed in #8, if things go wrong
with accessing Canvas, the user cannot check if she's entered the
correct access token or not.

Furthermore, the RepoBee configuration wizard does also use a text
input field when asking for the user's Git access token. Thus, this
change increases consistency as well.